+++ internal/helpers/traffic/pkg/app/init_network_node.go
@@ -3,6 +3,7 @@ package app
 import (
 	"github.com/number571/go-peer/pkg/network"
 	"github.com/number571/go-peer/pkg/network/conn"
+	hiddenlake "github.com/number571/hidden-lake"
 	"github.com/number571/hidden-lake/internal/helpers/traffic/internal/cache"
 	"github.com/number571/hidden-lake/internal/helpers/traffic/internal/handler"
 	"github.com/number571/hidden-lake/internal/helpers/traffic/internal/storage"
@@ -15,19 +16,19 @@ func (p *sApp) initNetworkNode(pStorage storage.IMessageStorage) {
 	p.fNode = network.NewNode(
 		network.NewSettings(&network.SSettings{
 			FAddress:      p.fConfig.GetAddress().GetTCP(),
-			FMaxConnects:  hls_settings.CNetworkMaxConns,
-			FReadTimeout:  hls_settings.CNetworkReadTimeout,
-			FWriteTimeout: hls_settings.CNetworkWriteTimeout,
+			FMaxConnects:  hiddenlake.GSettings.NetworkManager.ConnectsLimiter,
+			FReadTimeout:  hiddenlake.GSettings.GetReadTimeout(),
+			FWriteTimeout: hiddenlake.GSettings.GetWriteTimeout(),
 			FConnSettings: conn.NewSettings(&conn.SSettings{
 				FMessageSettings:       cfgSettings,
 				FLimitMessageSizeBytes: cfgSettings.GetMessageSizeBytes(),
-				FWaitReadTimeout:       hls_settings.CConnWaitReadTimeout,
-				FDialTimeout:           hls_settings.CConnDialTimeout,
-				FReadTimeout:           hls_settings.CNetworkReadTimeout,
-				FWriteTimeout:          hls_settings.CNetworkWriteTimeout,
+				FWaitReadTimeout:       hiddenlake.GSettings.GetWaitTimeout(),
+				FDialTimeout:           hiddenlake.GSettings.GetDialTimeout(),
+				FReadTimeout:           hiddenlake.GSettings.GetReadTimeout(),
+				FWriteTimeout:          hiddenlake.GSettings.GetWriteTimeout(),
 			}),
 		}),
-		cache.NewLRUCache(hls_settings.CNetworkQueueCapacity),
+		cache.NewLRUCache(hiddenlake.GSettings.NetworkManager.CacheHashesCap),
 	).HandleFunc(
 		hls_settings.CNetworkMask,
 		handler.HandleServiceTCP(p.fConfig, pStorage, p.fAnonLogger),