+++ internal/helpers/traffic/pkg/settings/settings.go
@@ -19,11 +19,6 @@ const (
 )
 
 const (
-	CDefaultMessagesCapacity = 0
-	CDefaultDatabaseEnabled  = false
-)
-
-const (
 	CDefaultTCPAddress  = "127.0.0.1:9581"
 	CDefaultHTTPAddress = "127.0.0.1:9582"
 )