+++ internal/service/pkg/app/init_conn_keeper.go
@@ -3,15 +3,14 @@ package app
 import (
 	"github.com/number571/go-peer/pkg/network"
 	"github.com/number571/go-peer/pkg/network/connkeeper"
-
-	pkg_settings "github.com/number571/hidden-lake/internal/service/pkg/settings"
+	hiddenlake "github.com/number571/hidden-lake"
 )
 
 func (p *sApp) initConnKeeper(pNode network.INode) {
 	p.fConnKeeper = connkeeper.NewConnKeeper(
 		connkeeper.NewSettings(&connkeeper.SSettings{
 			FConnections: func() []string { return p.fCfgW.GetConfig().GetConnections() },
-			FDuration:    pkg_settings.CConnKeeperDuration,
+			FDuration:    hiddenlake.GSettings.GetKeeperPeriod(),
 		}),
 		pNode,
 	)